+++ install.js
@@ -14,7 +14,7 @@ module.exports = {
           PIP_PROGRESS_BAR: "on"
         },
         message: [
-          "python -u scripts/launch_step.py \"Installing/upgrading pip, wheel, and setuptools\" -- python -m pip install --upgrade pip wheel setuptools --progress-bar on",
+          "python -u scripts/launch_step.py \"Installing/upgrading pip, wheel, and compatible setuptools\" -- python -m pip install --upgrade pip wheel setuptools==71.1.0 --progress-bar on",
           "python -m pip uninstall -y pyqt6-tools pyqt6-plugins qt6-tools qt6-applications",
           "python -u scripts/launch_step.py \"Installing BallonsTranslator requirements from requirements.txt\" -- uv pip install -r requirements.txt"
         ]